backend/requirements.txt
Django==4.2.5
psycopg2-binary==2.9.7
python-decouple==3.8
djangorestframework==3.14.0  
django-cors-headers==4.3.1